# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rental Mobil API",
        "description": "Membuat API sederhana untuk rental mobil",
        "contact": {
            "email": "surya.aw26@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/v1/car": {
            "get": {
                "tags": [
                    "car"
                ],
                "summary": "Get all data",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Failed",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/car/create": {
            "post": {
                "tags": [
                    "car"
                ],
                "summary": "Insert new car",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "brand_car": {
                                                "type": "string"
                                            },
                                            "car_model_year": {
                                                "type": "integer"
                                            },
                                            "color": {
                                                "type": "string"
                                            },
                                            "capacity": {
                                                "type": "integer"
                                            },
                                            "no_plate": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "brand_car": "Toyota",
                                    "car_model_year": "2016",
                                    "color": "blue",
                                    "capacity": "4",
                                    "no_plate": "AA1F28G"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Failed",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/car/{car_id}": {
            "get": {
                "tags": [
                    "car"
                ],
                "summary": "Get car by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Failed",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/car/edit/{car_id}": {
            "put": {
                "tags": [
                    "car"
                ],
                "summary": "Edit car",
                "operationId": "updatee",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "brand_car": {
                                                "type": "string"
                                            },
                                            "car_model_year": {
                                                "type": "integer"
                                            },
                                            "color": {
                                                "type": "string"
                                            },
                                            "capacity": {
                                                "type": "integer"
                                            },
                                            "no_plate": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "brand_car": "Toyota",
                                    "car_model_year": "2016",
                                    "color": "blue",
                                    "capacity": "4",
                                    "no_plate": "AA1F28G"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/car/delete/{car_id}": {
            "delete": {
                "tags": [
                    "car"
                ],
                "summary": "Delete car",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}